+++ frontend/src/router/index.ts
@@ -1,9 +1,12 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
 import Main from '../views/MainView.vue';
-import UserMenuView from '../views/UserMenuView.vue';
+import SettingView from '../views/SettingView.vue';
 import UserInfoView from '../views/user/InfoView.vue';
-import AuthService from '../service/AuthService'
+import UserCartView from '../views/user/CartView.vue';
+
+import AuthService from '../service/AuthService';
+
 
 Vue.use(VueRouter);
 
@@ -17,17 +20,25 @@ const routes = [
     },
   },
   {
-    path: '/user',
-    name: 'UserMenuView',
-    component: UserMenuView,
+    path: '/setting',
+    name: 'Setting',
+    component: SettingView,
+    meta: {
+      requireAuth: ture,
+    },
+  },
+  {
+    path: '/setting/info',
+    name: 'UserInfoView',
+    component: UserInfoView,
     meta: {
       requireAuth: true,
     },
   },
   {
-    path: '/user/info',
-    name: 'UserInfoView',
-    component: UserInfoView,
+    path: '/setting/cart',
+    name: 'UserCartView',
+    component: UserCartView,
     meta: {
       requireAuth: true,
     },
@@ -43,7 +54,7 @@ const router = new VueRouter({
 const authService = new AuthService();
 router.beforeEach(async (to, _, next) => {
   let user = await authService.getUser();
-  if (to.path === '/callback'){
+  if (to.path === '/callback') {
     await authService.siginCallback();
     next('/');
   }
